Agregación de confirmación para eliminación

+++ src/app/components/fiscal-positions/fiscal-position-list/fiscal-position-list/fiscal-position-list.component.ts
@@ -75,15 +75,19 @@ export class FiscalPositionListComponent implements AfterViewInit{
   }
 
   deleteFiscalPosition(id: number){
-    this.fiscalPositionService.deleteFiscalPosition(id).subscribe({
-      next: (res) => {
-        alert('Fiscal position deleted!');
-        this.getAllFiscalPositions();
-      },
-      error: () => {
-        alert('Error while deleting the fiscal position!')
-      }
-    })
+
+    if (confirm('Are you sure to delete the item with ID ' + id + '?')){
+      this.fiscalPositionService.deleteFiscalPosition(id).subscribe({
+        next: (res) => {
+          alert('Fiscal position deleted!');
+          this.getAllFiscalPositions();
+        },
+        error: () => {
+          alert('Error while deleting the fiscal position!')
+        }
+      })
+    }
+
   }
 
   openDialog() {
